report params:sampleRate in options iface.

+++ include/synthpod_app.h
@@ -77,7 +77,7 @@ struct _sp_app_system_sink_t {
 };
 
 struct _sp_app_driver_t {
-	uint32_t sample_rate;
+	float sample_rate;
 	uint32_t min_block_size;
 	uint32_t max_block_size;
 	uint32_t seq_size;
